requirements.txt
# FastAPI関連
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# RAG・LLM関連
# LlamaIndex: 0.9.0は存在しないため、0.10.0以降を使用
llama-index>=0.10.0,<0.15.0
openai>=1.0.0
tiktoken>=0.5.0

# データベース
sqlalchemy==2.0.23
aiosqlite==0.19.0

# 環境変数管理
python-dotenv==1.0.0

# その他
# Python 3.14対応のため、最新版のpydanticを使用
pydantic>=2.8.0
pydantic-settings>=2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# PDF生成（簡易版）
reportlab==4.0.7

# Wordファイル生成
python-docx==1.1.0

# Azure関連（Azure移行時に追加）
# azure-functions>=1.18.0  # Azure Functions用（Functionsアプリにデプロイする場合）
azure-storage-blob>=12.19.0  # Blob Storage用
azure-data-tables>=12.4.0  # Table Storage用
azure-identity>=1.15.0  # Managed Identity認証用

# 開発用（オプション）
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

